requirements.txt
discord-rewrite
asyncio
aiohttp
psutils